build!: add features for disabling sqlite-based implementation
Release-As: 0.7.0
