+++ src/repo.rs
@@ -1,25 +1,10 @@
-//! Code providing the interface for sequence repositories and the base `SeqRepo` implementation.
+//! Code providing the base `SeqRepo` implementation.
 
 use std::path::{Path, PathBuf};
 
 use crate::error::Error;
-use crate::{AliasDb, FastaDir, Namespace, Query};
-
-/// Trait describing the interface of a sequence repository.
-pub trait Interface {
-    /// Fetch part sequence given an alias.
-    fn fetch_sequence(&self, alias_or_seq_id: &AliasOrSeqId) -> Result<String, Error> {
-        self.fetch_sequence_part(alias_or_seq_id, None, None)
-    }
-
-    /// Fetch part sequence given an alias.
-    fn fetch_sequence_part(
-        &self,
-        alias_or_seq_id: &AliasOrSeqId,
-        begin: Option<usize>,
-        end: Option<usize>,
-    ) -> Result<String, Error>;
-}
+use crate::interface::Interface;
+use crate::{AliasDb, AliasOrSeqId, FastaDir, Namespace, Query};
 
 /// Provide (read-only) access to a `seqrepo` sequence repository.
 #[derive(Debug)]
@@ -32,14 +17,6 @@ pub struct SeqRepo {
     alias_db: AliasDb,
     /// The `FastaDir` to use.
     fasta_dir: FastaDir,
-}
-
-pub enum AliasOrSeqId {
-    Alias {
-        value: String,
-        namespace: Option<String>,
-    },
-    SeqId(String),
 }
 
 impl SeqRepo {
@@ -120,7 +97,7 @@ impl Interface for SeqRepo {
 
 #[cfg(test)]
 mod test {
-    use crate::{repo::Interface, AliasOrSeqId, SeqRepo};
+    use crate::{AliasOrSeqId, Interface, SeqRepo};
     use anyhow::Error;
 
     #[test]
